+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### pack, serve
 
 * Fixed `enact serve` fails with multiple ports by overriding resolvePort as devServer port.
+* Added `ENACT_PACK_ISOMORPHIC` as a global variable to use `hydrateRoot` instead of `createRoot` from app when isomorphic build.
 * Added `ignoreWarning` config to ignore warnings from SnapshotPlugin.
 * Updated webpack config to support `sass-loader` for opt-in support of SASS/SCSS files.